Use type sequence instantiation to declare values
This obscure language feature is documented in the following article on
dlang.org:

  https://dlang.org/articles/ctarguments.html

It does *not* appear to be documented in the language spec, which
mentions type sequences, but says only that they can be used in place of
argument lists of functions or templates.

Credit to Freddy A Cubas, author of the dub package 'tagged_union', for
his example of this technique in action.