Add test for stale pointers
If a pointer from the SumType's previous value is left in memory after a
new value is assigned, a conservative GC will fail to collect the memory
that pointer points to, causing a memory leak.

The existing behavior was correct, but the test is being left in to
guard against potential future regressions.

Thanks to John Colvin for pointing out this potential issue.