Make opAssign and constructor exception safe
If the assignment to storage fails after a new value has already been
assigned to tag, the SumType will be left in an invalid state. Assigning
to storage first ensures that this will never happen, since the
assignment to tag cannot fail.

Thanks to John Colvin for pointing out this issue.