Allow non-copyable types as members
Mutable rvalues are now moved into the SumType's storage, instead of
being copied, during construction and assignment. Non-mutable rvalues
and lvalues are still (by necessity) copied.

The use of 'move' breaks the test-case for types with invalid .init
values. Such types are impossible to detect with design-by-introspection
(since their assertions fail only at runtime), and were arguably an
abuse of the language to begin with, so throwing them under the bus
seems like an acceptable compromise.

The new compilation guard aroung 'toString' is needed because 'text',
like most of Phobos, is not designed to work with non-copyable types.