Remove useless template constraint
The intent of the removed constraint was to ensure that size_t.max could
never be a valid tag value. However, because the maximum length of
TypeArgs (and therefore Types) is size_t.max, the maximum possible tag
value (i.e., type index) was already size_t.max - 1, even without the
constraint.