Mark SumType.get as @system
The D language spec was recently updated with a precise definition of
what constitutes a "@safe interface". SumType.get does not meet this
definition. Therefore, it must be marked as @system, and the
responsibility for ensuring it is used safely must fall to its callers.

In practice, the only function that calls 'get' is 'matchImpl', and it
does so in a way that guarantees safety, so this change should have no
external impact.