Revert "Mark SumType.get as @system"
This reverts commit cbfdce0ee3987a1794dd217b384f8f329a35298d.

While SumType.get is not memory-safe in isolation, its safety depends on
the guarantees of SumType's public interface that

  1. A SumType's tag will always match the type of the stored value.
  2. A SumType cannot be assigned to in safe code if that assignment
     could cause unsafe aliasing.

Since neither of these depend upon the manner in which SumType.get is
called, it makes more sense to put the @trusted annotation on its
definition, rather than at the call site.

Note that this means @safe code in the sumtype module itself can violate
memory safety by breaking one of the above guarantees.