Don't rely on broken isAssignable behavior
Currently, isAssignable is true for non-copyable types, but should be
false. This will be fixed in an upcoming Phobos release, so add a
separate check for non-copyable types to avoid a regression when Phobos
is upgraded.

The symbol std.traits.isRvalueAssignable is currently private, but
expected to become public in a future Phobos release.

See also:
  - dlang issue 21210
  - dlang/phobos#7611