Use a custom 'This' type instead of std.variant's
Having a separate sumtype.This allows self-referential SumTypes and
Algebraics to be nested in each other without conflict. Making it a unit
type instead of an opaque type allows it to be used as a parameter and
return type in function types.

Attempting to import and use sumtype.This and std.variant.This in the
same scope will now cause errors due to ambiguity, so this is a breaking
change.

Fixes #49 on Github.