Relax @safe restrictions on opAssign
Overwriting a SumType's value with a pointer can be @safe so long as the
old value is known, at compile time, to be either (a) a non-pointer, or
(b) a pointer of the same type as the new value. This change allows such
assignments in @safe code, while still forbidding other assignments that
overwrite pointers.

See also:
  - https://issues.dlang.org/show_bug.cgi?id=22225
  - https://github.com/dlang/phobos/pull/8201

Fixes #67 on Github.