Add procedural API
This change introduces three new convenience functions that allow code
using SumType to be written in a more procedural style (as opposed to
the functional style of match).

* has!T checks whether a SumType contains a value of type T.
* get!T accesses a SumType's value if it has type T, or asserts if it
  does not.
* tryGet!T is like get!T, but throws an exception instead of asserting
  on failure.

To make room for get!T, the existing SumType.get method is renamed to
getByIndex.

Frontend requirement is bumped to 2.103 for __traits(fullyQualifiedName).

Phobos PR: https://github.com/dlang/phobos/pull/10650/